Retrieving the photos from an URL
This is for when the project requires retrieving the SVG icons from an URL instead of having them locally. I know it could be way better code but I don't have much more time to give it a thought. If someone wants to do it a better way you are more than welcome.